Changed the permission requeriments and moved protect to a concern
